+++ CHANGELOG.md
@@ -9,6 +9,7 @@ Dev-master
 * Enh: added externalVideo method into Content helper
 * Enh: added adBlockPosition method into Content helper
 * Enh: added slider method into Content helper
+* Enh: added additionalContent method into Content helper
 
 1.1.1 July 10, 2018
 --------------------